feat(model): multi parameter fitting
This took a long while as scipy.optimize(x,*p) takes a list of separate args
and not a vector list. I hardcoded for the case of SIR (two parameters).

Tomorrow I will extend it also to work for SIR-X and it is fine for the release,
but the code can be much improved without changing the functionality and extending
to an arbitrary length of parameters. I don't know how to do this at the moment
and creating a new minimizing function is not practical.