test(model): add initial tests
An __init__.py file was included into the tests folder, so
the tests can reach other modules.

Two tests are included into the file test_sir.py

1) test_sixdays(): Numerical convergence to a known value:
this is the  first test called test_sixdays() and tests
the accurate prediction of infected population on six days
for the Ealing borough.

2) test_expconv(): Numerical convergence to the analytical
solution in the simplified case of very large population
and few infected people (S->1).