feat(model): extend fit to SIR-X
This pull request enhances the model.fit() function in terms of its
accuracy and robustness:

In model.py, an special case for optimization is included for
the SIR-X model. One of the model parameters is the initial ratio
between infected and tested individuals, i_0/x_0. This is fitted
and require a slightly more careful treatment than traditional
parameter fitting. You can see the implementation on model.py fit

As I discovered this fifth parameter of sirx, I modified the sirx.py
class increasing the number of parameters to 5 and unpacking them
properly. Additionally, I included interesting quantities mentioned
in the paper of Maier et al (2020) such as the public containment
leverage P and the probability of quarantined Q. An effective r_0
and T_I are also defined. All of this was implemented as properties
in sirx.

I included two more attributes on SIR and SIR-X:
- name: this enables the fit() function to adjust the procedure
for the SIR or SIR-X (and will be helpful to extend for any other model)
- fit_input: Although SIR fits "I" directly from the number of
observed cases, SIR-X fits "X" to the observed cases and calculates I.
To specify to which index of the model outputs the input data should
be fitted, the fit_input is a property that the user can control.
By default, fit_input = 2 for SIR and fit_input = 4 for SIR-X