refactor(model): collect fit attributes
After a call with jia200x, we observed that for the release
2.0 we will need to freate a Fit and Predict classes later.
This commit points towards that direction packing all the
attributes that Fit uses into a dictionary fit_attr.

For now, it provides a good amount of transparency of the
dependencies. In the future, the fit_attr values will be
called with something like fit[attr], where fit would be
the new container