feat: SIR-X predict
Implements .predict function on SIR-X. n_X
and n_R are optional parameters. n_I is calculated
from the fitted infected_over_tested ratio. The
refactor on model.py was reverted as the differences
on .predict between the two models are relevant, and
in case of having more models these differences in
calling the function may have been even greater